split terminal.h into base header
